// services/react-frontend/src/library/common/types/measurementTypes.ts
export enum MeasurementTypes {
	CREATE_MEASUREMENT = '@@MEASUREMENT/CREATE_MEASUREMENT',
	CREATE_MEASUREMENT_SUCCESS = '@@MEASUREMENT/CREATE_MEASUREMENT_SUCCESS',
	EDIT_MEASUREMENT_NAME = '@@MEASUREMENT/EDIT_MEASUREMENT_NAME',
	DELETE_MEASUREMENT = '@@MEASUREMENT/DELETE_MEASUREMENT',
	REQUEST_MEASUREMENT_BY_ID = '@@MEASUREMENT/REQUEST_MEASUREMENT_BY_ID',
	REQUEST_MEASUREMENT_BY_ID_SUCCESS = '@@MEASUREMENT/REQUEST_MEASUREMENT_BY_ID_SUCCESS',
	EDIT_MEASUREMENT_NAME_SUCCESS = '@@MEASUREMENT/EDIT_MEASUREMENT_NAME_SUCCESS',
	DELETE_MEASUREMENT_SUCCESS = '@@MEASUREMENT/DELETE_MEASUREMENT_SUCCESS',
}